style : Network APIServer, HTTPMethod, HTTPHeader, ContentType 파일 분리
