+++ iOS/IssueTracker/IssueTracker/Network/APIServer.swift
@@ -1,0 +1,12 @@
+//
+//  APIServer.swift
+//  IssueTracker
+//
+//  Created by 송민관 on 2020/11/09.
+//
+
+import Foundation
+
+struct APIServer {
+    static let baseURL = "http://101.101.210.34:3000"
+}
